Change menu nodes to label type
Mouse input no longer works to select new game or load game.
New route node also is now an instance of save-button scene.
This scene type now has a new built-in script that defines a new signal, 'selected'.
This allows us to maintain more or less the same structure of code,
with the slight difference that now we don't force a mouse click input signal,
we use our own event for signaling when a file is selected.