Override hump's gamestate module with SWITCHER
SWITCHER is a new module that overrides hump's overriding of love's
callbacks. Now we no longer call Gamestate as a global, but we do so
of SWITCHER instead. SWITCHER makes hump's switching state change after
the function `love.draw` runs. This stops the screen from flickering
when pushing or popping states.