Preview static abilities in other abilities
Also change how static abilities work. Instead of expanding matched
commands, they now preprocess the parameters of matched commands. In the
case of Bloodthirst, for instance, it adds 1 to the "value" parameter of
damage_on_target commands.

This simplification is what enables the more detailed preview of effects
in other cards, so a Strike card now properly indicates it is receiving
a +1 damage.