Fix #1596: Add git submodules
Move all externals references to be tracked via git submodules. This way
each commit in this repository will be pointed to a particular commit of
each one of the external repositories, making it easier to ensure that
all compilations from the same commit have the same result.

When any of the external repositories change, any developer can do
"make update" and it will take the latest commit from each one. After
verifying that everything works, that same developer can push the
updates so they will be available for everyone in a new commit.